Upload indempotence bugfix
If you try to FTP deploy (upload) more than once, it can happen that the upload bails on you. Why? Because ftp-deploy attempts to (re-) create your directory structure, but can't, because it already exists. And if you use --force, all your code ends up piled up in a single directory. Splat.

I have modified the code so that it is idempotent (i.e. you can execute the same command more than once and get exactly the same result.) I simply nerfed the 'couldn't create dir' message from error to warning, and made sure that ftp-deploy attempts to change into directories that threw error messages when we tried to create them. 

To make it clear: An error when creating a directory does not imply directory nonexistence, and if ftp-deploy recognizes this, it becomes idempotent.